requirements.txt
addict==2.4.0
albumentations==1.3.1
descriptastorus==2.6.0
h5py==3.8.0
hydra-core==1.3.2
joblib==1.3.1
joblib==0.14.1
lmdb==1.4.1
matplotlib==3.7.2
numba==0.57.1
numpy==1.24.4
omegaconf==2.3.0
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==2.0.1
Pillow==9.2.0
Pillow==10.0.0
pymatgen==2023.7.20
PyYAML==6.0
PyYAML==6.0.1
rdkit==2023.3.2
scikit_learn==0.22.1
scikit_learn==1.1.3
scipy==1.9.3
setuptools==49.6.0
setuptools==67.8.0
torch==2.1.0
torchvision==0.13.0
tqdm==4.65.0
unicore==0.0.1
